Bug fix: multi-selection button functionality now works properly
